+++ crates/oxide/src/extractor/pre_processors/rust.rs
@@ -104,6 +104,12 @@ impl Rust {
 
                 b'[' => {
                     bracket_stack.push(cursor.curr());
+
+                    // Handle `p.flex[condition]`. If there is a `-` before it, it will likely be an
+                    // arbitrary value e.g. `text-[red]`
+                    if !matches!(cursor.prev(), b'-') {
+                        result[cursor.pos] = b' ';
+                    }
                 }
 
                 b']' if !bracket_stack.is_empty() => {
@@ -212,5 +218,21 @@ mod tests {
 
         let input = r#"html! { \x.px-4.text-black {  } }"#;
         Rust::test(input, r#"html! { \x px-4 text-black {  } }"#);
+    }
+
+    // https://github.com/tailwindlabs/tailwindcss/issues/20233
+    #[test]
+    fn test_maud_template_extraction_with_conditional_classes() {
+        let input = r#"
+            use maud::{html, Markup};
+
+            pub fn main() -> Markup {
+                html! {
+                    p.text-black[cuteness > 50] { "Squee!" }
+                }
+            }
+        "#;
+
+        Rust::test_extract_contains(input, vec!["text-black"]);
     }
 }